Fix Category A clippy warnings: Part 6 (imgproc remaining - 235 warnings)
Fixed all remaining Category A casting warnings in imgproc module:

Imgproc (235 warnings fixed, 7 files):
- drawing.rs: Fixed line, circle, rectangle, ellipse, polygon drawing (61 warnings)
- geometric.rs: Fixed resize, warp transformations, interpolation (58 warnings)
- morphology.rs: Fixed erosion, dilation, structuring elements (50 warnings)
- hough.rs: Fixed Hough line/circle detection, accumulator operations (41 warnings)
- edge.rs: Fixed Sobel, Laplacian, Scharr, Canny edge detection (23 warnings)
- histogram.rs: Fixed histogram equalization normalization (1 warning)
- filter.rs: Removed unnecessary cast (1 warning)

Patterns applied:
- Drawing: Dimension conversions (usize↔i32) with #[allow] for hot paths
- Geometric: Float↔pixel conversions with clamping
- Morphology: Coordinate calculations with bounds checking
- Hough: Accumulator indexing with precision loss annotations
- Edge: Gradient calculations with appropriate allows

All 230 tests passing. Imgproc module now 100% Category A compliant.

Progress: ~1,257/1,833 warnings fixed (68.6%)
Remaining: ~576 warnings (photo, video, features2d remaining, other modules)

+++ src/imgproc/filter.rs
@@ -47,7 +47,7 @@ pub async fn blur_async(src: &Mat, dst: &mut Mat, ksize: Size, use_gpu: bool) ->
         #[cfg(feature = "gpu")]
         {
             use crate::gpu::ops::box_blur_gpu_async;
-            match box_blur_gpu_async(src, dst, ksize.width as i32).await {
+            match box_blur_gpu_async(src, dst, ksize.width).await {
                 Ok(()) => return Ok(()),
                 Err(_) => {
                     // Fall through to CPU
